Refactor tests with sending email
